+++ src/SqlAdapters/SqlAdapter.cs
@@ -8,7 +8,7 @@ internal class SqlAdapter: ISqlAdapter {
 	/// <summary>
 	/// The SQL query used to fetch the identifier of the last inserted row.
 	/// </summary>
-	public string LastInsertIdQuery => "SELECT SCOPE_IDENTITY() AS Id";
+	public string LastInsertIdQuery => $"SELECT SCOPE_IDENTITY() AS {EscapeIdentifier("Id")}";
 
 	/// <summary>
 	/// Escapes the specified identifier.